Install Bazel via Ubuntu apt repository (#15)
While this no longer locks the version of Bazel down in the
configuration, it guarantees we're running against latest stable
release. While we lose reproducibility, we gain a fair bit of
convenience.

An alternative would be to run a periodic job that watches for new
versions of Bazel and generates, tests, and merges PRs that upgrade to
them.